+++ packages/U2Demo/src/package.ts
@@ -23,6 +23,7 @@ import '@datagrok-libraries/u2/css/list.css';
 import '@datagrok-libraries/u2/css/functions-browser.css';
 import '@datagrok-libraries/u2/css/grid.css';
 import '@datagrok-libraries/u2/css/icon-input.css';
+import '@datagrok-libraries/u2/css/function-input.css';
 import '@datagrok-libraries/u2/css/file.css';
 import '@datagrok-libraries/u2/css/progress.css';
 import '@datagrok-libraries/u2/css/spec.css';